modify class name and files
